Components default should be empty list, not empty string
Empty string results in a list with one item (`['']`) when
AnsibleModule casts str to list. This is not the intended
effect of omitting the `components` parameter, and results
in an error in the `release_fetch()` function, when it
tries to call `iocage fetch` with the `-F` flag and an
empty argument.